// backend/sanity.types.ts
export type Brand = {
  _id: string
  _type: 'brand'
  _createdAt: string
  _updatedAt: string
  _rev: string
  name?: string
  slug?: Slug
}
// ...
export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}